requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pytest==7.4.0
httpx==0.24.1 